IS-522 added attaching bonuses to homebrew items

+++ app/commands/bonuses_context/refresh_bonuses_command.rb
@@ -6,11 +6,11 @@ module BonusesContext
       config.messages.namespace = :bonus
 
       params do
-        required(:bonusable).filled(type?: ::Feat)
+        required(:bonusable).filled(type_included_in?: [::Feat, ::Item])
         optional(:bonuses).maybe(:array).each(:hash) do
           required(:id).filled(type_included_in?: [Integer, String])
-          required(:type).filled(:string)
-          required(:value)
+          optional(:type).filled(:string)
+          optional(:value)
         end
       end
     end
@@ -27,17 +27,13 @@ module BonusesContext
 
     def do_persist(input)
       input[:bonuses].each do |item|
-        attributes = item[:type] == 'static' ? { value: item[:value] } : { dynamic_value: item[:value] }
+        next if input[:ids_for_creating].exclude?(item[:id])
 
-        if input[:ids_for_creating].include?(item[:id])
-          Character::Bonus.create(attributes.merge(bonusable: input[:bonusable]))
-        else
-          bonus = Character::Bonus.find_by(id: item[:id])
-          bonus.update(attributes)
-        end
+        attributes = item[:type] == 'static' ? { value: item[:value] } : { dynamic_value: item[:value] }
+        Character::Bonus.create(attributes.merge(bonusable: input[:bonusable]))
       end
 
-      input[:bonusable].bonuses.where(id: input[:ids_for_removing]).delete_all
+      input[:bonusable].bonuses.where(id: input[:ids_for_removing]).delete_all if input[:ids_for_removing].any?
 
       { result: :ok }
     end
